Moved Manager Mock Build to Fixtures\HelperTrait, refs #11
