--- tests/Fixtures/JsonValueTrait.php
@@ -1,31 +1,0 @@
-<?php
-
-namespace Art4\JsonApiClient\Tests\Fixtures;
-
-/**
- * Link Object
- *
- * @see http://jsonapi.org/format/#document-links
- */
-trait JsonValueTrait
-{
-	/**
-	 * Json Values Provider
-	 *
-	 * @see http://json.org/
-	 */
-	public function jsonValuesProvider()
-	{
-		return array(
-			array(new \stdClass()),
-			array(array()),
-			array('string'),
-			array(456),
-			array(159.654),
-			array(-15E-3),
-			array(true),
-			array(false),
-			array(null),
-		);
-	}
-}
